Turn the proximity test into a devicelight test, since that is what it actually tests.
